+++ .changeset/nice-cups-act.md
@@ -1,0 +1,5 @@
+---
+"@beequeue/eslint-plugin": patch
+---
+
+Move project from TS to type annotated JS
