+++ .prettierignore
@@ -1,3 +1,3 @@
 package.json
-tsconfig.json
+jsconfig.json
 pnpm-lock.yaml
