+++ src/config/react.js
@@ -1,7 +1,6 @@
-import type { ESLintConfig } from "eslint-define-config"
+import { prettierStyleRules } from "../utils/prettier.js"
 
-import { prettierStyleRules } from "../utils/prettier"
-
+/** @type import("eslint-define-config").ESLintConfig */
 export const react = {
   extends: [
     /**
@@ -55,4 +54,4 @@ export const react = {
       },
     },
   ],
-} satisfies ESLintConfig
+}
